fix mpiio 2176, larger jobs are launched

+++ lbm_adios/lbm-adios/ds_adaptor.h
@@ -41,7 +41,7 @@ status_t ds_adaptor_init_client(int nprocs, int appid, MPI_Comm * pcomm, const c
  * @param elem_size element size
  * @param p_time_used timer
  */
-status_t get_common_buffer(uint8_t transport_minor, int timestep,int ndim, int bounds[6], int rank, char * var_name, void **p_buffer,size_t elem_size, double *p_time_used);
+status_t get_common_buffer(uint8_t transport_minor, int timestep,int ndim, uint64_t bounds[6], int rank, char * var_name, void **p_buffer,size_t elem_size, double *p_time_used);
  
 /*
  * @brief put data to dspaces/dimes
@@ -57,7 +57,7 @@ status_t get_common_buffer(uint8_t transport_minor, int timestep,int ndim, int b
  * @param p_time_used timer
  */
 
-status_t put_common_buffer(uint8_t transport_minor, int timestep, int ndim, int bounds[6], int rank, char * var_name, void **p_buffer,size_t elem_size, double *p_time_used);
+status_t put_common_buffer(uint8_t transport_minor, int timestep, int ndim, uint64_t bounds[6], int rank, char * var_name, void **p_buffer,size_t elem_size, double *p_time_used);
 
 //void get_common_buffer_unblocking(int timestep,int ndim, int bounds[6], int rank, MPI_Comm * p_gcomm,char * var_name, void **p_buffer,size_t elem_size, double *p_time_used);
 status_t ds_adaptor_flush_dimes(char * var_name, MPI_Comm comm, int nsteps);
